Update README.md
Added on to the list of available races and classes.
Added a section on how to use the code yourself. How to compile, build, and run the code. what resources I would recommend.
Added a section on functionality I would like to see implemented in case I come back to the program or others want to submit code themselves.
Removed unnecessary comments.
Removed sections of stuff I had in a to-do list, as I've done them.

This corresponding code upload is a big update to the program!